+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 0.1.0
+- added bulk exporting animation actions
+- made UI elements hide/show based on the export_method & selection state
+- got Source Tools Export functioning
+
 ## 0.0.0
 - initial commit